add a local-node logfile facility
If the riak env variable 'riak_local_logfile' is set, then all
events generated on this node will also be logged out to the
file named by the string in that variable.

Note that this is discouraged for production use, as the I/O
load generated by writing all of the events (there are many)
will compete with the core data storage I/O needs of Riak.