allow streaming of keylists and map/red over whole buckets
With this commit, the key-listing of buckets is performed in an
incremental streaming fashion without ever collecting the entire list
in one memory location.  This allows for the useful listing of much
larger buckets, especially if you use the streaming interface instead
of just the list_keys collector.

Also, the streaming keylist can be directed at a mapreduce job,
allowing mapreduce to be performed over a whole bucket in one go
instead of the more usual "targeted" mapreduce form.  Please note that
some (planned) optimizations in the mapreduce FSMs are needed before
this will have acceptable performance on very large buckets.

A demonstration, first that the existing list_keys behavior is
unchanged and then of the bucket mapreduce:

1: {ok,C} = riak:local_client()
-> {ok,{riak_client,'riak1@127.0.0.1',<<4,180,80,101>>}}
2: lists:foldl(fun(E, C) ->
                      O = riak_object:new(<<"testbuck">>, <<E:32>>, E),
                      C:put(O, 3),
                      C
               end,
               C,
               lists:seq(1, 1000))
-> {riak_client,'riak1@127.0.0.1',<<4,180,80,101>>}
3: {ok,L} = C:list_keys(<<"testbuck">>)
-> {ok,[...]}
4: length(L)
-> 1000
5: {ok,Res} =
       C:mapred_bucket(<<"testbuck">>,
                       [riak_mapreduce:map_object_value(false),
                        riak_mapreduce:reduce_sum(true)])
-> {ok,[500500]}

If you wish to consume the stream of keys directly, see the new
stream_list_keys functions in riak_client.